requirements.txt
Flask
flask_session
gunicorn
python-telegram-bot
spacy
langdetect
autocorrect
nltk
pyinflect
textblob
numpy
spacy_langdetect
spacy>=3.0.0,<4.0.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0.tar.gz#egg=en_core_web_sm
# click==8.0.3
# colorama==0.4.4
# cycler==0.10.0
# distlib==0.3.4
# et-xmlfile==1.1.0
# filelock==3.5.0
# Flask==2.0.2
# Flask-MySQL==1.5.2
# future==0.18.2
# gunicorn==20.1.0
# itsdangerous==2.0.1
# Jinja2==3.0.3
# joblib==1.0.1
# kiwisolver==1.3.1
# MarkupSafe==2.0.1
# matplotlib==3.4.1
# mysql-connector==2.2.9
# mysql-connector-python==8.0.27
# numexpr==2.7.3
# numpy==1.20.2
# opencv-contrib-python==4.5.2.52
# opencv-python==4.5.2.52
# openpyxl==3.0.9
# pandas==1.2.4
# pandastable==0.12.2.post1
# Pillow==8.2.0
# platformdirs==2.5.0
# protobuf==3.19.3
# PyMySQL==1.0.2
# pyparsing==2.4.7
# PySimpleGUI==4.43.0
# python-dateutil==2.8.1
# pytz==2021.1
# Sastrawi==1.0.1
# scikit-learn==0.24.2
# scipy==1.6.3
# seaborn==0.11.1
# six==1.15.0
# sklearn==0.0
# threadpoolctl==2.1.0
# virtualenv==20.13.1
# Werkzeug==2.0.2
# xlrd==2.0.1
